requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.40
psycopg2-binary>=2.9.9
jinja2>=3.1.4
python-multipart>=0.0.12
aiofiles>=24.1.0
markdown>=3.7
pytest>=8.3.3
httpx>=0.27.2
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
alembic>=1.13.0
email-validator>=2.1.0
